ENH: Updated directory structure and associated READMEs
Maintaining the mono-repo concept, but providing a directory
hierarchy that organizes the sub-projects into logical collections
that will help understand the general utility of each sub-project.

Also added a Data directory that provides a standard hierarchy for
accessing data from within the experiments and tests of this repo.
The data isn't included with this repo, but will mirror the
Data directory hierarchy in this project's shared gdrive folder (private).